Extract BinarTreeSet into reusable "suite" (#12)
